Поправила структуру, исправила нейминг
