+++ src/components/card-offers/card.tsx
