+++ src/components/offer-page/offer.tsx
