--- src/const.tsx
@@ -1,3 +1,0 @@
-export const Offers = {
-  CountOffers: 5,
-};
